pyproject.toml
[tool.poetry]
name = "ocr"
version = "0.1.0"
description = ""
authors = ["mzhadigerov <mzhadigerov@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
imutils = "^0.5.4"
argparse = "^1.4.0"
pytorch-pretrained-bert = "^0.6.2"
pyenchant = "^3.2.1"
nltk = "^3.6.2"
PassportEye = "^2.1.0"
pylev = "^1.4.0"
yapf = "^0.31.0"
isort = "^5.9.2"
click = "^8.0.1"
pytesseract = "^0.3.8"
pdf2image = "^1.16.0"
opencv-contrib-python = "^4.5.3"
Pillow = "^8.3.1"
tqdm = "^4.61.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
poethepoet = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
force-cuda11 = "python -m pip install torch==1.8.0+cu111 torchvision==0.9.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html"
